2011-07-04  Jeff Timanus  <twiz@chromium.org>
        Reviewed by Stephen White.

        Prevent calls to GrapicsContext3D::deleteTexture with id 0 in DrawingBuffer::clear().
        https://bugs.webkit.org/show_bug.cgi?id=63756

        * platform/graphics/gpu/DrawingBuffer.cpp:
        (WebCore::DrawingBuffer::clear):

Canonical link: https://commits.webkit.org/79635@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@90369 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/gpu/DrawingBuffer.cpp
@@ -58,36 +58,39 @@ void DrawingBuffer::clear()
 {
     if (!m_context)
         return;
-        
+
     m_context->makeContextCurrent();
-    m_context->deleteTexture(m_colorBuffer);
-    m_colorBuffer = 0;
-    
+
+    if (m_colorBuffer) {
+        m_context->deleteTexture(m_colorBuffer);
+        m_colorBuffer = 0;
+    }
+
     if (m_multisampleColorBuffer) {
         m_context->deleteRenderbuffer(m_multisampleColorBuffer);
         m_multisampleColorBuffer = 0;
     }
-    
+
     if (m_depthStencilBuffer) {
         m_context->deleteRenderbuffer(m_depthStencilBuffer);
         m_depthStencilBuffer = 0;
     }
-    
+
     if (m_depthBuffer) {
         m_context->deleteRenderbuffer(m_depthBuffer);
         m_depthBuffer = 0;
     }
-    
+
     if (m_stencilBuffer) {
         m_context->deleteRenderbuffer(m_stencilBuffer);
         m_stencilBuffer = 0;
     }
-    
+
     if (m_multisampleFBO) {
         m_context->deleteFramebuffer(m_multisampleFBO);
         m_multisampleFBO = 0;
     }
-        
+
     m_context->deleteFramebuffer(m_fbo);
     m_fbo = 0;
 }
